requirements.txt
Django==4.2.1
python-dotenv==1.0.0
django-widget-tweaks==1.4.12
django-multiupload==0.6.1
pandas==2.0.2
fitz==0.0.1.dev2
pymupdf==1.22.3
openai==0.27.7
asyncio==3.4.3
aiohttp==3.8.4
tiktoken==0.4.0
docx2txt==0.8
matplotlib==3.7.1 
plotly==5.14.1
sklearn==0.0.post5
scikit-learn==1.2.2
